refactor: 운영지역 패키지 구조 변경
